General Performance Tweaks (#31)
* Refactor GLDAStrategy::optimize


* Add APPROX_THRESHOLD


* Tweak invert signature


* Tweak profiler


* Tweak truncation logic


* Tweak Projector normalize logic


* Reserve vector memory in stratified_proportional_sample


* Allow for forest of 1 tree in Profiler


* Remove critical section from parallel for


* Use float instead of double


* Add Wno-ignored-attributes to Makevars


* Add tolerance to R tests


* Truncate determinant


* Add -Wa,-mbig-obj to debug build in Windows


---------

+++ core/src/Tree.hpp
@@ -65,7 +65,7 @@ namespace models {
       return !(*this == other);
     }
 
-    virtual double error_rate(const stats::SortedDataSpec<T, R> &data) const {
+    virtual float error_rate(const stats::SortedDataSpec<T, R> &data) const {
       auto [x, y, _classes] = data.unwrap();
       return stats::error_rate(predict(x), y);
     }
